Jlewi/analyze (#54)
Logging to support #7 

* Updated logging in Agent.Generate and Executor.Execute so we log the
information needed to determine when the human edits the cell

* To properly log protos in structured logging we need to use the zapr
proto plugin to generate MarshalLogObject functions for all the protos.
We can then use `zap.Object` to pass the protos as fields to logger
  * We need to enable logr to allow passing zap fields

* Add a logs subcommand and analyzer to process the logs and join the
log entries into traces and logs for the blocks

* Log call stack; fix #53; this was broken when we stopped using the
helper functions to build the logs. Now that we explicitly build the
logger we need to add the option to log the call stack.

* Altered the `SetupLogging` method in `app/cmd/assets.go` to accept a
`logToFile` boolean parameter, allowing conditional logging to files. We
only want the serve command to log to files. We don't need the other
commands to log to files


#### Analysis Package (`app/pkg/analyze`):
- Developed a new package, `analyze`, responsible for processing and
analyzing log files. This includes parsing raw log entries, organizing
them into traces and blocks, and producing structured summaries that
highlight the execution flow and outcomes.
- Implemented several utility types and methods (`LogEntry`, `Trace`,
`BlockLog`, etc.) to facilitate the analysis and organization of logs.

#### Configuration and Utility Changes:
- Extended the `app/pkg/config/config.go` to include methods for
retrieving both raw and processed log directories, supporting the new
logging mechanism.
- Added utility functions in `app/pkg/docs/outputs.go` to extract
specific metadata from logged blocks, such as execution exit codes.