Evaluation Results Data Explorer (#96)
Using a Google Sheet To View the evaluation results isn't working too
well. So we create our own viewer to display the results.

Create a connect-rpc service for listing the evaluation results

* This is the first step to building a viewer for the evaluation
results; we need a service to allow the client app to read the results.
* We decided to go with
[connect-rpc](https://connectrpc.com/docs/go/getting-started) because
this should simplify things by eliminating the need to use grpc and gprc
gateway in order to have a service that we can connect to from the
browser
* Furthermore, hopefully we can use the connect-rpc generated clients
even when using WASM

* Fix the go_package option in most of the proto files. We seem to be
missing a directory "foyle" in the path. This doesn't appear to matter
until we start using connect-rpc and then the imports are broken unless
we fix the package

Refactor our existing go-app to better support multiple pages. Each page
might render the right hand side differently.

* Add a status bar to display the GOAPP_VERSION; this makes it easier to
ensure we have picked up any changes.

The app is pretty ugly. I will try to fix the styling in a subsequent
PR.