Continuous log processing (#139)
This is the first half of continuous training. This implements the
logging piece of the processing.

There is a tech note describing it but the actual implementation made a
couple simplifications as a first pass. Notably, we aren't using a
durable queue for tracking blocklog events. Therefore, its possible if
Foyle crashes or on shutdown some blocks will not have been processed.
We can fix this later.

Create LockingDB as a way to deal with multiple writers trying to modify
the same row concurrently. This ensures concurrent operations are
sequenced and that the writes don't end up overwriting each other.

`foyle logs analyze` is now a null op because logs analysis runs
automatically in the background

Related to #84