Publish protos to the BSR; upgrade buf to v2 config (#172)
Fix #171 - Publish the SDKs and Protos to the BSR
* This will allow us to import the SDKs from the BSR into the vscode
extension (related to #167)

Finish upgrading the buf configurations to v2; use the buf CLI tool

Remove the typescript rules from `buf.gen.yaml`. Since we are publishing
to the BSR we no longer need to generate the typescript SDK because we
will just import it form the registry.