Update Foyle to filter exposition and return single code cell - Do well on single code cell generation (#194)
The purpose of this PR is to try to create a UX with GhostCells that
isn't so distracting and confusing as to cause people to disable it.

* Surfacing markdown suggestions is very confusing because we don't have
a visual affordance for ghost markdown cells #168
* More generally rendering multiple suggested ghost cells is visually
confusing

So this PR simplifies things by post processing the results to only
return a code block and even then return at most one code block. We
should nail this first before trying to expand to handle more
complicated scenarios.

I edited the prompt as well to try to make it more clear. I don't have
enough of an eval dataset to eval it and see if its improving things.


Related to #167
Related to #170