Allow Example Files To Be Overwritten (#210)
* An example file can be overwritten.
* A block could be executed multiple times. Each time a block is
executed it could trigger feedback and learning. In this case, the
learner would be invoked and we'd want to use the latest value of the
cell to learn from.
* Prior to https://github.com/jlewi/monogo/pull/22 FileHelper.NewWriter
would throw an error if the file already exists. That PR changed the
semantics to truncate the file if it already exists.
* So by upgrading monogo we should allow overwriting of examples if they
already exist.

Fix #205