Fix segfault in learner.
When learner is initialized at program startup we want to initialize an empty
array because learned examples might arrive during program execution.

If we have a nil matrix to store them we will get a segfault.

Ensuring a matrix is always initialized at startup is one way to solve this.