Render OpenAI request/response as HTML (#265)
This PR adds support for rendering of OpenAI requests/responses as HTML
(fix #259). This functionality was already supported for Anthropic.

I added an LLM sub-command to support rendering the data as HTML. This
makes it reusable and not closely tied to the GetLLMLogs RPC.

Fix HTML rendering of our prompts to properly handle the XML tags in our
prompt. We need to escape them so they aren't treated as raw HTML tags
which don't render.

Remove the obsolete logs subcommand.

Fix #259