Speedup Eval - Instrument With Otel (#302)
# Speedup Evaluation

Per #301, our experiments were really slow. We were observing times of
approximately 30s to 60s to process individual examples. With the new
dataset of 424 examples it would take ~7hours to run an experiment;
significantly impeding iteration.

This PR instruments evaluation with OTEL so that we can see the
bottlenecks in the code. Below are some graphs.

The graph below shows the P95 latency of processing evaluation examples.
We can see its about 30s.
<img width="1148" alt="experiment_traces_p95_latency"
src="https://github.com/user-attachments/assets/7f3d8aba-f3ed-4358-9dbc-4cb7fd292172">

Below is a heatmap showing the duration of waitForBlockLog
<img width="1207" alt="experiment_traces_heatmap_wait_for_block_log"
src="https://github.com/user-attachments/assets/7213fb6e-669d-4b46-b7e1-450154652962">

We can see that its about 20s-30s so it accounts for a large portion of
the duration.

A big source of the latency is the Analyzer uses a rate limiting queue
for reprocessing the logs. The max delay is 30s so we are probably
reprocessing the logs and 30s intervals. So to speed it up we make the
delay configurable so we can use a shorter delay during experiments.

Here is an updated heatmap of the duration of wait for block log.
<img width="1190" alt="Screen Shot 2024-10-15 at 4 34 40 PM"
src="https://github.com/user-attachments/assets/76d75baf-ae45-4c32-8c80-7bf6f1b3387e">

16:25 is when it started running with a maxDelaySeconds of 1s. We can
see this drops the latency down significantly from about 30s to ~6s