Improve Handling Of Long Outputs (#328)
# Problem

Cell outputs can be very long. For example, if we run a query (gcloud,
SQL, etc...) the output could be very verbose. This output could eat up
the entire context allocated for the input document. As a result, we
might not have sufficiently meaningful context to prompt the model.

There was another bug in our doc tailer. We were applying character
limits to the rendered markdown. We were imposing this by tailing the
lines. This could produce invalid markdown. For example, we might end up
truncating the document in the middle of a code block so we wouldn't
have the opening triple quotes for the code block. We might also include
the output of the code block without including the code that it is
output for.

# Solution

First, we impose character limits in a way that is aware of cell
boundaries. We move truncation into the Block to Markdown conversion.
The conversion now takes the maximum length for the output string. The
conversion routine then figures out how much to allocate to the contents
of the cell and its outputs. This allows truncation to happen in a way
that can respect cell boundaries.

Second, if we truncate the code block or output we output a string
indicating that the output was truncated. We want the model to know that
output was truncated. We update our prompt to tell the LLM to look for
truncated output and to potentially deal with this by running commands
that will provide less verbose output.



Fix #299