Fix Panic: Write After Handler Closed (#332)
I think we get this issue (#319) because we check whether the context is
cancelled before generating an AI response and then try to return the
response. However, in between the initial check and actually having the
response to send the context could have been cancelled.

So check if the context is done. before sending the response.

Fix #319