Add logging to confirm that stdout isn't empty when triggering on cell execution (#333)
* The purpose of this PR is to add logging to confirm that #309 is
fixed; i.e. that when triggering on cell execution for interactive (and
non interactive cells) the output is included in the request

* This PR adds logging that checks if the selected cell has non-empty
stdout if its empty we log a message. The presence of these log messages
thus indicates cell execution triggered completion that included no
stdout.

* This will include false positives because in certain cases the command
may not produce any output.

* With this logging I was unable to reproduce any instances where
completion was triggered by cell execution but stdout wasn't included.

* This PR fixes a bug in the mimetype for stdout; there was an extra
space. I don't think that actually impacted processing because that
constant was only used in unittests before this PR.

* fix #309